\name{all_pkg_stat_snapshot}
\alias{all_pkg_stat_snapshot}
\title{
The complete table of dependency heaviness for all CRAN/Bioconductor packages
}
\description{
The complete table of dependency heaviness for all CRAN/Bioconductor packages
}
\usage{
all_pkg_stat_snapshot()
}
\value{
The columns are self-explanatory from the column names.
}
\examples{
# There is no example
NULL
}
