\name{dependency_database}
\alias{dependency_database}
\title{
The database of package dependency analysis
}
\description{
The database of package dependency analysis
}
\usage{
dependency_database()
}
\details{
The analyis is based on all 22076 CRAN/Bioconductor packages retrieved on 2022-06-08.
}
\examples{
if(interactive()) {
    dependency_database()
}
}
