\name{dependency_website}
\alias{dependency_website}
\title{
Open the online website of package dependency analysis
}
\description{
Open the online website of package dependency analysis
}
\usage{
dependency_website()
}
\details{
The analyis is based on all CRAN/Bioconductor packages retrieved on 2021-10-28.
}
\examples{
# There is no example
NULL

}
