\name{gini_index}
\alias{gini_index}
\title{
Gini index on the heaviness
}
\description{
Gini index on the heaviness
}
\usage{
gini_index(x, a = 2)
}
\arguments{

  \item{x}{An object returned by \code{\link{pkgndep}}.}
  \item{a}{A constant add to heaviness.}

}
\examples{
# There is no example
NULL
}
