\name{heaviness_from_upstream}
\alias{heaviness_from_upstream}
\title{
Heaviness from all upstream packages
}
\description{
Heaviness from all upstream packages
}
\usage{
heaviness_from_upstream(package)
}
\arguments{

  \item{package}{A package name.}

}
\value{
A named vector.
}
\examples{
# There is no example
NULL

}
