\name{is_parent}
\alias{is_parent}
\title{
Test the parent-child relationship
}
\description{
Test the parent-child relationship
}
\usage{
is_parent(parent, child, ...)
}
\arguments{

  \item{parent}{A vector of package names.}
  \item{child}{A single package name.}
  \item{...}{Pass to \code{\link{parent_dependency}}.}

}
\value{
A logical vector.
}
\examples{
# There is no example
NULL

}
