\name{is_upstream}
\alias{is_upstream}
\title{
Test upstream - downstream relationship
}
\description{
Test upstream - downstream relationship
}
\usage{
is_upstream(upstream, package, ...)
}
\arguments{

  \item{upstream}{A vector of package names.}
  \item{package}{A single package name.}
  \item{...}{Pass to \code{\link{upstream_dependency}}.}

}
\value{
A logical vector.
}
\examples{
# There is no example
NULL

}
