\name{load_from_heaviness_db}
\alias{load_from_heaviness_db}
\title{
Load pre-computed objects
}
\description{
Load pre-computed objects
}
\usage{
load_from_heaviness_db(file)
}
\arguments{

  \item{file}{File name.}

}
\details{
The pathway of the file can be set via \code{\link{pkgndep_opt}}$db_file_template.

Internally used.
}
\examples{
# There is no example
NULL

}
