\name{load_heaviness_timeline}
\alias{load_heaviness_timeline}
\title{
Load heaviness statistics at all time points
}
\description{
Load heaviness statistics at all time points
}
\usage{
load_heaviness_timeline()
}
\details{
Used internally.
}
\value{
A list of data frames.
}
\examples{
# There is no example
NULL

}
