\name{pkgndep}
\alias{pkgndep}
\title{
Package dependency analysis
}
\description{
Package dependency analysis
}
\usage{
pkgndep(package, load = FALSE, verbose = TRUE)
}
\arguments{

  \item{package}{Package name. The value should be 1. a CRAN/Bioconductor package, 2. an installed package, 3. a path of a local package, 4. URL of a GitHub repository.}
  \item{load}{Check which other packages are loaded into R session (directly or indirectly) when loading \code{pkg}. }
  \item{verbose}{Whether to show messages.}

}
\details{
The package database for dependency analysis is always directly retrieved from CRAN/Bioconductor.
}
\value{
A \code{pkgndep} object.
}
\examples{
\dontrun{
x = pkgndep("ComplexHeatmap")
}
# The `x` variable generated by `pkgndep()` is already saved in this package.
x = readRDS(system.file("extdata", "ComplexHeatmap_dep.rds", package = "pkgndep"))
x
plot(x)
}
