\name{plot.pkgndep}
\alias{plot.pkgndep}
\title{
Make the dependency heatmap
}
\description{
Make the dependency heatmap
}
\usage{
\method{plot}{pkgndep}(x, ...)
}
\arguments{

  \item{x}{An object from \code{\link{pkgndep}}.}
  \item{...}{Other arguments.}

}
\details{
Please use \code{\link{dependency_heatmap}} instead.
}
\value{
A vector of two numeric values (in inches) that correspond to the width and height of the plot.

}
\examples{
# See examples in `pkgndep()`.

}
