\name{print.heaviness_db_info}
\alias{print.heaviness_db_info}
\title{
Print the heaviness_db_info object
}
\description{
Print the heaviness_db_info object
}
\usage{
\method{print}{heaviness_db_info}(x, ...)
}
\arguments{

  \item{x}{A \code{heaviness_db_info} object.}
  \item{...}{Other arguments}

}
\value{
No value is returned.
}
\examples{
heaviness_db
}
