\name{print.pkgndep}
\alias{print.pkgndep}
\title{
Print method
}
\description{
Print method
}
\usage{
\method{print}{pkgndep}(x, ...)
}
\arguments{

  \item{x}{An object from \code{\link{pkgndep}}.}
  \item{...}{Other arguments.}

}
\value{
No value is returned.
}
\examples{
# See examples in `pkgndep()`.
}
