\name{unavailable_packages}
\alias{unavailable_packages}
\title{
Unavailable packages
}
\description{
Unavailable packages
}
\usage{
unavailable_packages(x)
}
\arguments{

  \item{x}{An object from \code{\link{pkgndep}}.}

}
\details{
It lists the packages that are not installed in the "Suggests" field.
}
\value{
A vector of package names.
}
\examples{
# There is no example
NULL

}
