\name{which_imported}
\alias{which_imported}
\title{
Which packages are listed in Depends/Imports
}
\description{
Which packages are listed in Depends/Imports
}
\usage{
which_imported(x)
}
\arguments{

  \item{x}{The object from \code{\link{pkgndep}}.}

}
\value{
A vector of logical values.
}
\examples{
# There is no example
NULL
}
