\name{check_pkg}
\alias{check_pkg}
\title{
Check whether a package is available
}
\description{
Check whether a package is available
}
\usage{
check_pkg(pkg, bioc = FALSE)
}
\arguments{

  \item{pkg}{The name of the package.}
  \item{bioc}{Whether it is a Bioconductor package.}

}
\details{
If the package is not avaiable, it prints messages to guide users to install the package.
}
\examples{
# There is no example
NULL

}
