\name{dependency_report}
\alias{dependency_report}
\title{
HTML report for package dependency heaviness analysis
}
\description{
HTML report for package dependency heaviness analysis
}
\usage{
dependency_report(...)
}
\arguments{

  \item{...}{Pass to \code{\link{heaviness_report}}.}

}
\details{
It is the same as \code{\link{heaviness_report}}.
}
\examples{
# There is no example
NULL

}
