\name{get_all_functions_imported_to_children}
\alias{get_all_functions_imported_to_children}
\title{
Get functions that are imported to child packages
}
\description{
Get functions that are imported to child packages
}
\usage{
get_all_functions_imported_to_children(package)
}
\arguments{

  \item{package}{Package name.}

}
\value{
It returns a list of names of functions that are imported to every of its child packages.
}
\examples{
# There is no example
NULL

}
