\name{plot.pkgndep}
\alias{plot.pkgndep}
\title{
Make the dependency heatmap
}
\description{
Make the dependency heatmap
}
\usage{
\method{plot}{pkgndep}(x, ...)
}
\arguments{

  \item{x}{An object from \code{\link{pkgndep}}.}
  \item{...}{Other arguments.}

}
\details{
Please use \code{\link{dependency_heatmap}} instead.
}
\examples{
# There is no example
NULL

}
