\name{required_dependency_packages}
\alias{required_dependency_packages}
\title{
Required dependency packages
}
\description{
Required dependency packages
}
\usage{
required_dependency_packages(x, all = FALSE)
}
\arguments{

  \item{x}{An object from \code{\link{pkgndep}}.}
  \item{all}{Whether to include the packages required if also including packages from "Suggests"/"Enhances" field.}

}
\value{
A vector of package names.
}
\examples{
# There is no example
NULL

}
