\name{unavailable_pkg}
\alias{unavailable_pkg}
\title{
Unavailable packages
}
\description{
Unavailable packages
}
\usage{
unavailable_pkg(x)
}
\arguments{

  \item{x}{The object from \code{\link{pkgndep}}.}

}
\details{
It lists the packages that are not installed in the "Suggests" field.
}
\value{
A vector of package names.
}
\examples{
# There is no example
NULL

}
